changed doubles to floats in second raycaster
tasks to do:
-use dot product for collision w/non-orthogonal walls
-2d zbuffer for transparency in walls
-sprites
-multiplayer

+++ src/com/dezzy/skorp3/skorp3D/raycast/core/Vector2.java
@@ -1,11 +1,11 @@
 package com.dezzy.skorp3.skorp3D.raycast.core;
 
 public class Vector2 {
-	public double x;
-	public double y;
-	public double length = -1;
+	public float x;
+	public float y;
+	public float length = -1;
 	
-	public Vector2(double _x, double _y) {
+	public Vector2(float _x, float _y) {
 		x = _x;
 		y = _y;
 	}
@@ -15,15 +15,15 @@ public class Vector2 {
 	}
 	
 	public void updateLength() {
-		length = Math.sqrt((x * x) + (y * y));
+		length = (float)Math.sqrt((x * x) + (y * y));
 	}
 	
 	public Vector2 subtract(Vector2 v) {
 		return new Vector2(x - v.x, y - v.y);
 	}
 	
-	public static double distance(Vector2 _v0, Vector2 _v1) {
-		return Math.sqrt(((_v0.x - _v1.x) * (_v0.x - _v1.x)) + ((_v0.y - _v1.y) * (_v0.y - _v1.y)));
+	public static float distance(Vector2 _v0, Vector2 _v1) {
+		return (float) Math.sqrt(((_v0.x - _v1.x) * (_v0.x - _v1.x)) + ((_v0.y - _v1.y) * (_v0.y - _v1.y)));
 	}
 	
 	@Override
